requirements.txt
djangorestframework==3.15.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
pandas==2.2.3